requirements.txt
geopy==1.11.0
httplib2==0.9.1
python-instagram==1.3.2
simplejson==3.8.0
six==1.9.0
SQLAlchemy==1.0.8